+++ eslint.config.js
@@ -1,11 +1,18 @@
 import { eslint_config } from 'airier'
 
-export default eslint_config.map( config => config.rules
-    ? {
-        ...config,
-        rules: {
-            ...config.rules,
-            indent: `off`,
-        },
-    }
-    : config )
+export default [
+    {
+        ignores: [
+            `.browser-profile/**`,
+        ],
+    },
+    ...eslint_config.map( config => config.rules
+        ? {
+            ...config,
+            rules: {
+                ...config.rules,
+                indent: `off`,
+            },
+        }
+        : config ),
+]